backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
PyGithub==2.1.1
requests==2.31.0
python-dotenv==1.0.0
matplotlib>=3.9.0
plotly==5.18.0
kaleido==0.2.1
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.11.0
pydantic==2.5.3
python-multipart==0.0.6
